polish the code that marks new as unresolved
Previously, because `mark_new_as_unresolved` was hooked on `save_post`, it would cause edits to fire the action and thus re-change the state to new. This resulted in edits to a post to act as if this was a new post.

This fixes the problem by hooking into post transitions instead.